requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
mammoth>=1.4.1
altair>=4.2.0
google-generativeai>=1.0.0
mammoth
